+++ examples/shared_cell.rs
@@ -1,3 +1,5 @@
+use core::pin::pin;
+
 use shared_cell::SharedCell;
 
 struct Context {
@@ -6,9 +8,9 @@ struct Context {
 
 fn main() {
     let mut context = Context { stuff: 42 };
-    let mut context = SharedCell::new(&mut context);
+    let mut context = pin!(SharedCell::new(&mut context));
 
-    context.with(|cx| {
+    context.as_mut().with(|cx| {
         println!("Before: {}", cx.stuff);
         cx.stuff += 1;
         println!("After: {}", cx.stuff);