+++ src/shared_cell.rs
@@ -1,6 +1,8 @@
 use core::{
     cell::Cell,
     fmt::{Debug, Formatter, Result},
+    marker::PhantomPinned,
+    pin::Pin,
 };
 
 /// Shared cell type
@@ -10,7 +12,7 @@ use core::{
 /// ```
 #[doc = include_str!("../examples/shared_cell.rs")]
 /// ```
-pub struct SharedCell<'a, T: ?Sized>(&'a Cell<T>);
+pub struct SharedCell<'a, T: ?Sized>(&'a Cell<T>, PhantomPinned);
 
 impl<T: ?Sized> Debug for SharedCell<'_, T> {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
@@ -23,7 +25,7 @@ impl<T: ?Sized> Debug for SharedCell<'_, T> {
 impl<'a, T: ?Sized> SharedCell<'a, T> {
     /// Create a new [`SharedCell`]
     pub fn new(value: &'a mut T) -> Self {
-        Self(Cell::from_mut(value))
+        Self(Cell::from_mut(value), PhantomPinned)
     }
 
     /// Duplicate the [`SharedCell`].
@@ -36,16 +38,25 @@ impl<'a, T: ?Sized> SharedCell<'a, T> {
     ///    ability to resume execution of an asynchronous task that holds onto
     ///    another [`SharedCell`].
     pub unsafe fn duplicate(&mut self) -> Self {
-        Self(self.0)
+        Self(self.0, PhantomPinned)
     }
 
-    /// Acquires a mutable reference to the cell's interior value.
-    pub fn with<R>(&mut self, f: impl FnOnce(&mut T) -> R) -> R {
+    /// Acquire a mutable reference to the cell's interior value.
+    pub fn with<R>(self: Pin<&mut Self>, f: impl FnOnce(&mut T) -> R) -> R {
         // SAFETY: By isolating the `SharedCell` to one instance per scope, we
         // prevent reëntrant calls to `with()`.
         //
         // SAFETY: Cannot yield to code that could call `with()` due to safety
         // invariant on `duplicate()`
         unsafe { f(&mut *self.0.as_ptr()) }
+    }
+
+    /// Get a mutable refernce to the internal data.
+    ///
+    /// # Safety
+    ///
+    ///  - There must be no duplicated instances of [`SharedCell`].
+    pub unsafe fn into_inner(self) -> &'a mut T {
+        &mut *self.0.as_ptr()
     }
 }